Added new PowerShell script
The purpose of printparserwin2008-R2-GetPrintLogsAndSum.ps1 is to
connect to a remote print server running Windows Server 2008 R2, and get
the print logs. After acquiring those logs, it takes all print jobs by a
user and sums the total amount of pages they have printed.
This script is split between two functions (one of which may be removed
later), the first function gets all users and there total pages printed.
The second function only looks for those users who have printed jobs
over the $magicNumber and totals there total amount of pages that were a
part of a job that went over the $magicNumber. This is a bit silly on my
part.
In the future, I hope to change the second function to take the first
set of results (all users) and just filter out those who have a total
page count greater than or equal to the $magicNumber. This may or may
not be any easy task.